frontend/requirements.txt
flet>=0.21.2
requests>=2.31.0
python-dotenv>=1.0.0
python-dateutil>=2.8.2
pytz>=2023.3
httpcore>=1.0.2
httpx>=0.25.0
dnspython>=2.4.2
python-jose[cryptography]>=3.3.0
